pyproject.toml
[tool.poetry]
name = "project frontend"
version = "0.1.0"
description = ""
authors = ["Eldar Mametov <e.mametov@innopolis.university>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
